Modifiy ImportNewD2Document on the Repository class to use d2-config vs. referring to it as a profile since it is NOT a profile.
D2Config class now sets on the D2Document object to create a new D2 Document with the given config. Some work needs to be done to do profile/property config lookups to do pre-validation checks on object create/save events.

Not implemented yet, but adding a new option for debugging the JsonDotNetSerializer that allows you to ouput the text of the stream before Deserializing (help you ensure your C# object matches what is in the JSON text).

I hope to get the RestError object in this week to handle bubbling up errors from the server side to the client (if desired) or at least in the client/consumer log.